refactor: render contact menu option on mobile only, make language apear inside the mobile nav menu
