+++ src/components/transtions/index.ts
@@ -1,1 +1,2 @@
 export * from './page-transition';
+export * from './language-selector';
